refactor(#97): vehicles/types — bulletproof-react segmentation
git mv:
- projection/vehicleGqlShaped.ts → types/vehicleGqlShaped.ts
- xml/Vehicle.ts, VehicleModel.ts, *-mapping.ts → types/

ESLint exemption paths for netex-ts-gen output updated to types/.
Intra-feature import paths follow.
