+++ src/data/vehicles/types/Vehicle-mapping.ts
