+++ src/data/vehicles/types/Vehicle.ts
