+++ src/data/vehicles/types/VehicleModel-mapping.ts
