+++ src/data/vehicles/types/VehicleModel.ts
