+++ src/data/vehicles/types/vehicleGqlShaped.ts
