+++ src/data/vehicles/vehicleFormState.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { edit, hydrate, initialFormState, isDirty, type FormState } from './vehicleFormState';
 import { BLANK_FORM, MISSING_MODEL } from './vehicleFormDefaults';
-import type { Vehicle } from './xml/Vehicle';
+import type { Vehicle } from './types/Vehicle';
 import type { VehicleEditFormValue } from './VehicleEditForm';
 
 const aVehicle: Partial<Vehicle> = {
