+++ src/data/vehicles/xml/buildVehiclePairXml.ts
@@ -1,6 +1,6 @@
 import { buildPublicationDeliveryXml } from '../../netex/publicationDeliveryXml';
-import { vehicleToXmlShape } from './Vehicle-mapping';
-import { vehicleModelToXmlShape } from './VehicleModel-mapping';
+import { vehicleToXmlShape } from '../types/Vehicle-mapping';
+import { vehicleModelToXmlShape } from '../types/VehicleModel-mapping';
 import type { VehicleEditFormValue } from '../VehicleEditForm';
 
 export function buildVehiclePairXml(form: VehicleEditFormValue): string {
